up-structs test
- Add `FrameKind::Scheam`, implement Frame for Schema, and make sure
stacks always have at least one frame (in fact, Stack's constructor now
takes a frame).  This gets rid of RustStack's `root()` hack, it now
returns as many supers as there are frames, not + 1.
- Add `Frame::body()`, implement for all 4 kinds.
- When generating functions, use proper stack (ie one that has
the Function on it) when writing params/results. This was already
done for sub-functions, so it simplifies things.
- Introduce `Stack::lookup_struct()`, which, well, looks up a struct.
It returns a `RelativePath`, which should be easy to codegen with
in non-Rust languages.
- Change rust ast_ext `RustBaseType` to `GeneratesRust` so it can
be implemented for `RelativePath` as well.
- output: Change CommaList to List, so it can generate other kinds of
separator-separated output, ie. not only (a, b, c) but also a::b::c.
Also, add 'Brackets::None' variant.
- Change all tests to use a schema named 'test' to make it easier to
rename them.
- Change all tests to import `services::test::*`, to avoid one layer
of namespacing.
- Add `up-structs` test, that uses a struct defined in an upper
namespace.
- Test runner: allow running single test